# gawk-admin/openapi.yaml
...
openapi: 3.1.1
info:
  title: gawk-admin API
  version: 1.3.0 # x-release-please-version
  summary: Moderation and fleet-operations API for a gawk deployment.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
  description: |
    The HTTP API of `gawk-admin`, the moderation portal of a
    [gawk](https://github.com/Tuhis/gawk) deployment. The portal SPA uses it,
    and external integrations such as bots use the same API.

    ## Authentication

    Every operation except `GET /api/v1/openapi.json` requires a bearer JWT
    issued by the deployment's OIDC provider:

    ```
    Authorization: Bearer <token>
    ```

    Each operation lists the role it requires under `x-gawk-roles`. A valid
    token without that role gets `403`. There are no cookies and no
    server-side session.

    The served copy of this document names the role claim value this
    deployment is configured with (`-operator-role`). The repository copy uses
    the symbolic name `operator`.

    Service identities use the client-credentials grant against the same
    provider. `docs/self-hosting.md` §9.8 has a worked example.

    The routes that need no token: `GET /auth/config` returns the issuer,
    client ID and audience a browser client needs; `GET /api/v1/openapi.json`
    returns this document; `GET /api/v1/asyncapi.json` and
    `GET /api/v1/schemas/events/{name}` return the event contract.

    ## Compatibility

...
    treat unknown enum values as unknown rather than as errors.

...

...

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/rooms/{name}:
    delete:
      operationId: deleteRoom
      tags: [rooms]
      summary: Delete a room
      description: |
        Deletes a room of either kind. Every relay pod acts on the deleted
        custom resource; for a dynamic room that closes the sessions in it.
      x-gawk-roles: [operator]
      x-gawk-requires: rooms
      parameters:
        - name: name
          in: path
          required: true
          description: The room code.
          schema: { type: string }
      responses:
        '204':
          description: The room was deleted.
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/Internal' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        An access token issued by the deployment's OIDC provider. Its role
        claim must include the role the operation lists under `x-gawk-roles`.
  responses:
...
    Unauthorized:
      description: The bearer token is missing or invalid.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error:
              code: unauthorized
              message: a valid bearer token is required
    Forbidden:
      description: The token lacks the required role.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error:
              code: forbidden
              message: this account does not hold the operator role
    TooManyRequests:
      description: Too many rejected tokens from this address.
      headers:
        Retry-After:
          description: Seconds to wait before retrying.
          schema: { type: integer, format: int32 }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error:
              code: rate_limited
              message: too many invalid credentials; slow down
    NotFound:
      description: No such object, or the feature is disabled.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error:
              code: not_found
              message: no such endpoint
...
    Internal:
      description: Unexpected server error.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error:
              code: internal
              message: internal error
...
  schemas:
    ErrorCode:
      type: string
      description: |
        Machine-readable error code. The enum only grows; treat unknown values
        as unknown. The first four are issued by the authentication layer and
        can occur on any operation that requires a role.
      enum:
        # From the authentication layer, on any operation with a role.
        - unauthorized
        - idp_unavailable
        - forbidden
        - rate_limited
        # From the handlers.
        - bad_request
        - not_found
        - duplicate_active
        - duplicate_name
        - source_immutable
        - internal
        - unavailable
        - invalid_target
        - ban_not_active
        - room_exists
        - room_not_static
        - room_not_dynamic
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code: { $ref: '#/components/schemas/ErrorCode' }
        message:
          type: string
          description: Human-readable explanation, for display only.
    Error:
      type: object
      description: The error envelope every failing operation returns.
      required: [error]
      properties:
        error: { $ref: '#/components/schemas/ErrorBody' }
...